+++ src/config.rs
@@ -9,6 +9,7 @@ use gumdrop::Options;
 use serde::{Deserialize, Serialize};
 use simplelog::*;
 use std::path::PathBuf;
+use std::str::FromStr;
 
 use crate::logger::GooseLogFormat;
 use crate::metrics::GooseCoordinatedOmissionMitigation;
@@ -155,6 +156,12 @@ pub struct GooseConfiguration {
     /// Sets how many times to run scenarios then exit
     #[options(no_short)]
     pub iterations: usize,
+    /// Limits load test to only specified scenarios
+    #[options(no_short, meta = "\"SCENARIO\"")]
+    pub scenarios: Scenarios,
+    /// Lists all scenarios and exits
+    #[options(no_short)]
+    pub scenarios_list: bool,
     /// Doesn't enable telnet Controller
     #[options(no_short)]
     pub no_telnet: bool,
@@ -221,6 +228,44 @@ pub struct GooseConfiguration {
     pub manager_port: u16,
 }
 
+/// Optionally defines a subset of active Scenarios to run during a load test.
+#[derive(Options, Default, Debug, Clone, Serialize, Deserialize)]
+pub struct Scenarios {
+    pub active: Vec<String>,
+}
+/// Implement [`FromStr`] to convert `"foo,bar"` comma separated string to a vector of strings.
+impl FromStr for Scenarios {
+    type Err = GooseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        // Convert string into a vector of string.
+        let mut active: Vec<String> = Vec::new();
+        // Multiple Scenarios can be defined as a comma separated list.
+        let lines = s.split(',');
+        for line in lines {
+            // Ignore white space an case.
+            let scenario = line.trim().to_lowercase();
+            // Valid scenario names are alphanumeric only.
+            if scenario.chars().all(char::is_alphanumeric) {
+                active.push(scenario);
+            } else {
+                // Logger isn't initialized yet, provide helpful debug output.
+                eprintln!("ERROR: invalid `configuration.scenarios` value: '{}'", line);
+                eprintln!("  Expected format: --scenarios \"{{one}},{{two}},{{three}}\"");
+                eprintln!("    {{one}}, {{two}}, {{three}}, etc must be alphanumeric");
+                eprintln!("    To view valid scenario names invoke `--scenarios-list`");
+                return Err(GooseError::InvalidOption {
+                    option: "`configuration.scenarios".to_string(),
+                    value: line.to_string(),
+                    detail: "invalid `configuration.scenarios` value.".to_string(),
+                });
+            }
+        }
+        // The listed scenarios are only valid if the logic gets this far.
+        Ok(Scenarios { active })
+    }
+}
+
 /// Optional default values for Goose run-time options.
 ///
 /// These custom defaults can be configured using [`GooseDefaultType::set_default()`].
@@ -240,6 +285,8 @@ pub(crate) struct GooseDefaults {
     pub test_plan: Option<TestPlan>,
     /// An optional default test plan.
     pub iterations: Option<usize>,
+    /// Optional default scenarios.
+    pub scenarios: Option<Scenarios>,
     /// An optional default log level.
     pub log_level: Option<u8>,
     /// An optional default for the goose log file name.
@@ -354,6 +401,8 @@ pub enum GooseDefault {
     TestPlan,
     /// An optional default number of iterations to run scenarios then exit.
     Iterations,
+    /// Optional default list of scenarios to run.
+    Scenarios,
     /// An optional default log level.
     LogLevel,
     /// An optional default for the log file name.
@@ -479,21 +528,22 @@ pub enum GooseDefault {
 ///
 /// The following run-time options can be configured with a custom default using a
 /// borrowed string slice ([`&str`]):
-///  - [`GooseDefault::Host`]
-///  - [`GooseDefault::Timeout`]
-///  - [`GooseDefault::HatchRate`]
-///  - [`GooseDefault::ReportFile`]
-///  - [`GooseDefault::GooseLog`]
-///  - [`GooseDefault::RequestLog`]
-///  - [`GooseDefault::TransactionLog`]
-///  - [`GooseDefault::ScenarioLog`]
-///  - [`GooseDefault::ErrorLog`]
 ///  - [`GooseDefault::DebugLog`]
-///  - [`GooseDefault::TestPlan`]
-///  - [`GooseDefault::TelnetHost`]
-///  - [`GooseDefault::WebSocketHost`]
+///  - [`GooseDefault::ErrorLog`]
+///  - [`GooseDefault::GooseLog`]
+///  - [`GooseDefault::HatchRate`]
+///  - [`GooseDefault::Host`]
 ///  - [`GooseDefault::ManagerBindHost`]
 ///  - [`GooseDefault::ManagerHost`]
+///  - [`GooseDefault::ReportFile`]
+///  - [`GooseDefault::RequestLog`]
+///  - [`GooseDefault::ScenarioLog`]
+///  - [`GooseDefault::Scenarios`]
+///  - [`GooseDefault::TelnetHost`]
+///  - [`GooseDefault::TestPlan`]
+///  - [`GooseDefault::Timeout`]
+///  - [`GooseDefault::TransactionLog`]
+///  - [`GooseDefault::WebSocketHost`]
 ///
 /// The following run-time options can be configured with a custom default using a
 /// [`usize`] integer:
@@ -572,8 +622,10 @@ impl GooseDefaultType<&str> for GooseAttack {
     fn set_default(mut self, key: GooseDefault, value: &str) -> Result<Box<Self>, GooseError> {
         match key {
             // Set valid defaults.
+            GooseDefault::DebugLog => self.defaults.debug_log = Some(value.to_string()),
+            GooseDefault::ErrorLog => self.defaults.error_log = Some(value.to_string()),
+            GooseDefault::GooseLog => self.defaults.goose_log = Some(value.to_string()),
             GooseDefault::HatchRate => self.defaults.hatch_rate = Some(value.to_string()),
-            GooseDefault::Timeout => self.defaults.timeout = Some(value.to_string()),
             GooseDefault::Host => {
                 self.defaults.host = if value.is_empty() {
                     None
@@ -581,22 +633,23 @@ impl GooseDefaultType<&str> for GooseAttack {
                     Some(value.to_string())
                 }
             }
-            GooseDefault::GooseLog => self.defaults.goose_log = Some(value.to_string()),
-            GooseDefault::ReportFile => self.defaults.report_file = Some(value.to_string()),
-            GooseDefault::RequestLog => self.defaults.request_log = Some(value.to_string()),
-            GooseDefault::TransactionLog => self.defaults.transaction_log = Some(value.to_string()),
-            GooseDefault::ScenarioLog => self.defaults.scenario_log = Some(value.to_string()),
-            GooseDefault::ErrorLog => self.defaults.error_log = Some(value.to_string()),
-            GooseDefault::DebugLog => self.defaults.debug_log = Some(value.to_string()),
-            GooseDefault::TelnetHost => self.defaults.telnet_host = Some(value.to_string()),
-            GooseDefault::WebSocketHost => self.defaults.websocket_host = Some(value.to_string()),
             GooseDefault::ManagerBindHost => {
                 self.defaults.manager_bind_host = Some(value.to_string())
             }
             GooseDefault::ManagerHost => self.defaults.manager_host = Some(value.to_string()),
+            GooseDefault::ReportFile => self.defaults.report_file = Some(value.to_string()),
+            GooseDefault::RequestLog => self.defaults.request_log = Some(value.to_string()),
+            GooseDefault::ScenarioLog => self.defaults.scenario_log = Some(value.to_string()),
+            GooseDefault::Scenarios => {
+                self.defaults.scenarios = Some(value.parse::<Scenarios>().unwrap())
+            }
+            GooseDefault::TelnetHost => self.defaults.telnet_host = Some(value.to_string()),
             GooseDefault::TestPlan => {
                 self.defaults.test_plan = Some(value.parse::<TestPlan>().unwrap())
             }
+            GooseDefault::Timeout => self.defaults.timeout = Some(value.to_string()),
+            GooseDefault::TransactionLog => self.defaults.transaction_log = Some(value.to_string()),
+            GooseDefault::WebSocketHost => self.defaults.websocket_host = Some(value.to_string()),
             // Otherwise display a helpful and explicit error.
             GooseDefault::Users
             | GooseDefault::StartupTime
@@ -695,21 +748,22 @@ impl GooseDefaultType<usize> for GooseAttack {
             GooseDefault::ManagerBindPort => self.defaults.manager_bind_port = Some(value as u16),
             GooseDefault::ManagerPort => self.defaults.manager_port = Some(value as u16),
             // Otherwise display a helpful and explicit error.
-            GooseDefault::Host
-            | GooseDefault::TestPlan
-            | GooseDefault::HatchRate
-            | GooseDefault::Timeout
+            GooseDefault::DebugLog
+            | GooseDefault::ErrorLog
             | GooseDefault::GooseLog
+            | GooseDefault::HatchRate
+            | GooseDefault::Host
+            | GooseDefault::ManagerBindHost
+            | GooseDefault::ManagerHost
             | GooseDefault::ReportFile
             | GooseDefault::RequestLog
-            | GooseDefault::TransactionLog
             | GooseDefault::ScenarioLog
-            | GooseDefault::ErrorLog
-            | GooseDefault::DebugLog
+            | GooseDefault::Scenarios
             | GooseDefault::TelnetHost
-            | GooseDefault::WebSocketHost
-            | GooseDefault::ManagerBindHost
-            | GooseDefault::ManagerHost => {
+            | GooseDefault::TestPlan
+            | GooseDefault::Timeout
+            | GooseDefault::TransactionLog
+            | GooseDefault::WebSocketHost => {
                 return Err(GooseError::InvalidOption {
                     option: format!("GooseDefault::{:?}", key),
                     value: format!("{}", value),
@@ -799,20 +853,22 @@ impl GooseDefaultType<bool> for GooseAttack {
             GooseDefault::NoGranularData => self.defaults.no_granular_report = Some(value),
             GooseDefault::Worker => self.defaults.worker = Some(value),
             // Otherwise display a helpful and explicit error.
-            GooseDefault::Host
-            | GooseDefault::TestPlan
+            GooseDefault::DebugLog
+            | GooseDefault::ErrorLog
             | GooseDefault::GooseLog
+            | GooseDefault::HatchRate
+            | GooseDefault::Host
+            | GooseDefault::ManagerBindHost
+            | GooseDefault::ManagerHost
             | GooseDefault::ReportFile
             | GooseDefault::RequestLog
-            | GooseDefault::TransactionLog
             | GooseDefault::ScenarioLog
-            | GooseDefault::RunningMetrics
-            | GooseDefault::ErrorLog
-            | GooseDefault::DebugLog
+            | GooseDefault::Scenarios
             | GooseDefault::TelnetHost
-            | GooseDefault::WebSocketHost
-            | GooseDefault::ManagerBindHost
-            | GooseDefault::ManagerHost => {
+            | GooseDefault::TestPlan
+            | GooseDefault::Timeout
+            | GooseDefault::TransactionLog
+            | GooseDefault::WebSocketHost => {
                 return Err(GooseError::InvalidOption {
                     option: format!("GooseDefault::{:?}", key),
                     value: format!("{}", value),
@@ -823,10 +879,9 @@ impl GooseDefaultType<bool> for GooseAttack {
                 })
             }
             GooseDefault::Users
-            | GooseDefault::HatchRate
-            | GooseDefault::Timeout
             | GooseDefault::StartupTime
             | GooseDefault::RunTime
+            | GooseDefault::RunningMetrics
             | GooseDefault::Iterations
             | GooseDefault::LogLevel
             | GooseDefault::Quiet
@@ -911,20 +966,23 @@ impl GooseDefaultType<GooseCoordinatedOmissionMitigation> for GooseAttack {
                     ),
                 })
             }
-            GooseDefault::Host
-            | GooseDefault::TestPlan
+            // Otherwise display a helpful and explicit error.
+            GooseDefault::DebugLog
+            | GooseDefault::ErrorLog
             | GooseDefault::GooseLog
+            | GooseDefault::HatchRate
+            | GooseDefault::Host
+            | GooseDefault::ManagerBindHost
+            | GooseDefault::ManagerHost
             | GooseDefault::ReportFile
             | GooseDefault::RequestLog
-            | GooseDefault::TransactionLog
             | GooseDefault::ScenarioLog
-            | GooseDefault::RunningMetrics
-            | GooseDefault::ErrorLog
-            | GooseDefault::DebugLog
+            | GooseDefault::Scenarios
             | GooseDefault::TelnetHost
-            | GooseDefault::WebSocketHost
-            | GooseDefault::ManagerBindHost
-            | GooseDefault::ManagerHost => {
+            | GooseDefault::TestPlan
+            | GooseDefault::Timeout
+            | GooseDefault::TransactionLog
+            | GooseDefault::WebSocketHost => {
                 return Err(GooseError::InvalidOption {
                     option: format!("GooseDefault::{:?}", key),
                     value: format!("{:?}", value),
@@ -935,10 +993,9 @@ impl GooseDefaultType<GooseCoordinatedOmissionMitigation> for GooseAttack {
                 })
             }
             GooseDefault::Users
-            | GooseDefault::HatchRate
-            | GooseDefault::Timeout
             | GooseDefault::StartupTime
             | GooseDefault::RunTime
+            | GooseDefault::RunningMetrics
             | GooseDefault::Iterations
             | GooseDefault::LogLevel
             | GooseDefault::Quiet
@@ -1017,20 +1074,23 @@ impl GooseDefaultType<GooseLogFormat> for GooseAttack {
                     ),
                 })
             }
-            GooseDefault::Host
-            | GooseDefault::TestPlan
+            // Otherwise display a helpful and explicit error.
+            GooseDefault::DebugLog
+            | GooseDefault::ErrorLog
             | GooseDefault::GooseLog
+            | GooseDefault::HatchRate
+            | GooseDefault::Host
+            | GooseDefault::ManagerBindHost
+            | GooseDefault::ManagerHost
             | GooseDefault::ReportFile
             | GooseDefault::RequestLog
-            | GooseDefault::TransactionLog
             | GooseDefault::ScenarioLog
-            | GooseDefault::RunningMetrics
-            | GooseDefault::ErrorLog
-            | GooseDefault::DebugLog
+            | GooseDefault::Scenarios
             | GooseDefault::TelnetHost
-            | GooseDefault::WebSocketHost
-            | GooseDefault::ManagerBindHost
-            | GooseDefault::ManagerHost => {
+            | GooseDefault::TestPlan
+            | GooseDefault::Timeout
+            | GooseDefault::TransactionLog
+            | GooseDefault::WebSocketHost => {
                 return Err(GooseError::InvalidOption {
                     option: format!("GooseDefault::{:?}", key),
                     value: format!("{:?}", value),
@@ -1041,10 +1101,9 @@ impl GooseDefaultType<GooseLogFormat> for GooseAttack {
                 })
             }
             GooseDefault::Users
-            | GooseDefault::HatchRate
-            | GooseDefault::Timeout
             | GooseDefault::StartupTime
             | GooseDefault::RunTime
+            | GooseDefault::RunningMetrics
             | GooseDefault::Iterations
             | GooseDefault::LogLevel
             | GooseDefault::Quiet
@@ -1073,7 +1132,6 @@ impl GooseDefaultType<GooseLogFormat> for GooseAttack {
                         key, value
                     ),
                 })
-
             }
         }
         Ok(Box::new(self))
@@ -1115,7 +1173,7 @@ impl GooseConfigure<usize> for GooseConfiguration {
     }
 }
 impl GooseConfigure<u16> for GooseConfiguration {
-    /// Use [`GooseValue`] to set a [`usize`] value.
+    /// Use [`GooseValue`] to set a [`u16`] value.
     fn get_value(&self, values: Vec<GooseValue<u16>>) -> Option<u16> {
         for value in values {
             if let Some(v) = value.value {
@@ -1246,6 +1304,24 @@ impl GooseConfigure<GooseCoordinatedOmissionMitigation> for GooseConfiguration {
         &self,
         values: Vec<GooseValue<GooseCoordinatedOmissionMitigation>>,
     ) -> Option<GooseCoordinatedOmissionMitigation> {
+        for value in values {
+            if let Some(v) = value.value {
+                if value.filter {
+                    continue;
+                } else {
+                    if !value.message.is_empty() {
+                        info!("{} = {:?}", value.message, v)
+                    }
+                    return Some(v);
+                }
+            }
+        }
+        None
+    }
+}
+impl GooseConfigure<Scenarios> for GooseConfiguration {
+    /// Use [`GooseValue`] to set a [`Scenarios`] value.
+    fn get_value(&self, values: Vec<GooseValue<Scenarios>>) -> Option<Scenarios> {
         for value in values {
             if let Some(v) = value.value {
                 if value.filter {
@@ -1669,6 +1745,24 @@ impl GooseConfiguration {
                 },
             ])
             .unwrap_or(0);
+
+        // Configure `scenarios`.
+        self.scenarios = self
+            .get_value(vec![
+                // Use --scenarios if set.
+                GooseValue {
+                    value: Some(self.scenarios.clone()),
+                    filter: self.scenarios.active.is_empty(),
+                    message: "scenarios",
+                },
+                // Use GooseDefault if not already set and not Worker.
+                GooseValue {
+                    value: defaults.scenarios.clone(),
+                    filter: defaults.scenarios.is_none() || self.worker,
+                    message: "scenarios",
+                },
+            ])
+            .unwrap_or_else(|| Scenarios { active: Vec::new() });
 
         // Configure `no_debug_body`.
         self.no_debug_body = self