+++ src/lib.rs
@@ -63,9 +63,7 @@ use lazy_static::lazy_static;
 use nng::Socket;
 use rand::seq::SliceRandom;
 use rand::thread_rng;
-use std::collections::hash_map::DefaultHasher;
-use std::collections::BTreeMap;
-use std::collections::HashSet;
+use std::collections::{hash_map::DefaultHasher, BTreeMap, HashSet};
 use std::hash::{Hash, Hasher};
 use std::sync::{
     atomic::{AtomicUsize, Ordering},
@@ -383,6 +381,8 @@ pub struct GooseAttack {
     test_stop_transaction: Option<Transaction>,
     /// A vector containing one copy of each Scenario defined by this load test.
     scenarios: Vec<Scenario>,
+    /// A set of all registered scenario names.
+    scenario_machine_names: HashSet<String>,
     /// A weighted vector containing a GooseUser object for each GooseUser that will run during this load test.
     weighted_users: Vec<GooseUser>,
     /// A weighted vector containing a lightweight GaggleUser object that is sent to all Workers if running in Gaggle mode.
@@ -426,6 +426,7 @@ impl GooseAttack {
             test_start_transaction: None,
             test_stop_transaction: None,
             scenarios: Vec::new(),
+            scenario_machine_names: HashSet::new(),
             weighted_users: Vec::new(),
             weighted_gaggle_users: Vec::new(),
             defaults: GooseDefaults::default(),
@@ -462,6 +463,7 @@ impl GooseAttack {
             test_start_transaction: None,
             test_stop_transaction: None,
             scenarios: Vec::new(),
+            scenario_machine_names: HashSet::new(),
             weighted_users: Vec::new(),
             weighted_gaggle_users: Vec::new(),
             defaults: GooseDefaults::default(),
@@ -575,6 +577,22 @@ impl GooseAttack {
     /// ```
     pub fn register_scenario(mut self, mut scenario: Scenario) -> Self {
         scenario.scenarios_index = self.scenarios.len();
+        // Machine names must be unique. If this machine name has already been seen, add an
+        // integer at the end to differentiate.
+        let mut conflicts: u32 = 0;
+        let mut machine_name = scenario.machine_name.to_string();
+        // Inserting into the scenario_machine_names hashset will fail if this is name was
+        // already seen.
+        while !self.scenario_machine_names.insert(machine_name) {
+            // For each conflict increase the counter and try again.
+            conflicts += 1;
+            machine_name = format!("{}_{}", scenario.machine_name, conflicts);
+        }
+        // If there was a conflict, also update the scenario itself.
+        if conflicts > 0 {
+            scenario.machine_name = format!("{}_{}", scenario.machine_name, conflicts);
+        }
+        // Finally, register the scenario.
         self.scenarios.push(scenario);
         self
     }
@@ -646,6 +664,24 @@ impl GooseAttack {
         self
     }
 
+    /// Internal helper to determine if the scenario is currently active.
+    fn scenario_is_active(&self, scenario: &Scenario) -> bool {
+        // All scenarios are enabled by default.
+        if self.configuration.scenarios.active.is_empty() {
+            true
+        // Returns true or false depending on if the machine name is included in the
+        // configured `--scenarios`.
+        } else {
+            for active in &self.configuration.scenarios.active {
+                if scenario.machine_name.contains(active) {
+                    return true;
+                }
+            }
+            // No matches found, this scenario is not active.
+            false
+        }
+    }
+
     /// Use configured GooseScheduler to build out a properly weighted list of
     /// [`Scenario`](./goose/struct.Scenario.html)s to be assigned to
     /// [`GooseUser`](./goose/struct.GooseUser.html)s
@@ -655,13 +691,15 @@ impl GooseAttack {
         let mut u: usize = 0;
         let mut v: usize;
         for scenario in &self.scenarios {
-            if u == 0 {
-                u = scenario.weight;
-            } else {
-                v = scenario.weight;
-                trace!("calculating greatest common denominator of {} and {}", u, v);
-                u = util::gcd(u, v);
-                trace!("inner gcd: {}", u);
+            if self.scenario_is_active(scenario) {
+                if u == 0 {
+                    u = scenario.weight;
+                } else {
+                    v = scenario.weight;
+                    trace!("calculating greatest common denominator of {} and {}", u, v);
+                    u = util::gcd(u, v);
+                    trace!("inner gcd: {}", u);
+                }
             }
         }
         // 'u' will always be the greatest common divisor
@@ -671,18 +709,20 @@ impl GooseAttack {
         let mut available_scenarios = Vec::with_capacity(self.scenarios.len());
         let mut total_scenarios = 0;
         for (index, scenario) in self.scenarios.iter().enumerate() {
-            // divide by greatest common divisor so vector is as short as possible
-            let weight = scenario.weight / u;
-            trace!(
-                "{}: {} has weight of {} (reduced with gcd to {})",
-                index,
-                scenario.name,
-                scenario.weight,
-                weight
-            );
-            let weighted_sets = vec![index; weight];
-            total_scenarios += weight;
-            available_scenarios.push(weighted_sets);
+            if self.scenario_is_active(scenario) {
+                // divide by greatest common divisor so vector is as short as possible
+                let weight = scenario.weight / u;
+                trace!(
+                    "{}: {} has weight of {} (reduced with gcd to {})",
+                    index,
+                    scenario.name,
+                    scenario.weight,
+                    weight
+                );
+                let weighted_sets = vec![index; weight];
+                total_scenarios += weight;
+                available_scenarios.push(weighted_sets);
+            }
         }
 
         info!(
@@ -848,6 +888,19 @@ impl GooseAttack {
         }
     }
 
+    // Display all scenarios (sorted by machine name).
+    fn print_scenarios(&self) {
+        let mut scenarios = BTreeMap::new();
+        println!("Scenarios:");
+        for scenario in &self.scenarios {
+            scenarios.insert(scenario.machine_name.clone(), scenario.clone());
+        }
+        // Display sorted by machine_name.
+        for (key, scenario) in scenarios {
+            println!(r#" - {}: ("{}")"#, key, scenario.name);
+        }
+    }
+
     /// Execute the [`GooseAttack`](./struct.GooseAttack.html) load test.
     ///
     /// # Example
@@ -917,6 +970,27 @@ impl GooseAttack {
 
         // Validate GooseConfiguration.
         self.configuration.validate()?;
+
+        // Display scenarios, then exit.
+        if self.configuration.scenarios_list {
+            self.print_scenarios();
+            std::process::exit(0);
+        }
+
+        // At least one scenario must be active.
+        let mut active_scenario: bool = false;
+        for scenario in &self.scenarios {
+            if self.scenario_is_active(scenario) {
+                active_scenario = true;
+                break;
+            }
+        }
+        if !active_scenario {
+            self.print_scenarios();
+            return Err(GooseError::NoScenarios {
+                detail: "No scenarios are enabled.".to_string(),
+            });
+        }
 
         // Build TestPlan.
         self.test_plan = TestPlan::build(&self.configuration);