+++ tests/redirect.rs
@@ -332,7 +332,12 @@ async fn run_standalone_test(test_type: TestType) {
 
     // Run the Goose Attack.
     common::run_load_test(
-        common::build_load_test(configuration, &get_transactions(&test_type), None, None),
+        common::build_load_test(
+            configuration,
+            vec![get_transactions(&test_type)],
+            None,
+            None,
+        ),
         None,
     )
     .await;
@@ -361,7 +366,7 @@ async fn run_gaggle_test(test_type: TestType) {
     let worker_handles = common::launch_gaggle_workers(EXPECT_WORKERS, || {
         common::build_load_test(
             worker_configuration.clone(),
-            &get_transactions(&test_type),
+            vec![get_transactions(&test_type)],
             None,
             None,
         )
@@ -374,7 +379,7 @@ async fn run_gaggle_test(test_type: TestType) {
     // Build the load test for the Workers.
     let manager_goose_attack = common::build_load_test(
         manager_configuration,
-        &get_transactions(&test_type),
+        vec![get_transactions(&test_type)],
         None,
         None,
     );